Initial commit
Had problems with old repo (my fault) so had to recreate it.
Project already in advanced stage, but looking ahead to remake it using
Chromium Embedded for better printing (on windows, at least)...
Will try to keep things more engine-agnostic so it can be easely ported
to Qt 5 when it launches, if it offers a decent print support in
QtWebKit.